Move `Image.transposed()` to `ImageProtocol`
It makes `ImageSlice.transpozed()` available.
